ppr/requirements.txt
wheel==0.37.1
fastapi==0.85.1
uvicorn[standard]==0.18.3
content-size-limit-asgi==0.1.5
PyYAML==6.0
setuptools-scm==7.0.5
